+++ java-spring-boot-back-end-app/calling-card/src/main/java/com/example/calling_card/repositories/SavedCardsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-    public interface SavedCardsRepository extends JpaRepository <SavedCards, Integer> {
+    public interface SavedCardsRepository extends JpaRepository <SavedCards, Long> {
     }
